# api/docs/swagger.yaml
definitions:
  controllers.QrValidationRequest:
    properties:
      payload_json:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/public/battery/{bpan}:
    get:
      consumes:
      - application/json
      description: Returns only public fields for a battery (no authentication required)
      parameters:
      - description: BPAN
        in: path
        name: bpan
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Public battery data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Battery not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get public battery information
      tags:
      - public
  /api/v1/public/qr/validate:
    post:
      consumes:
      - application/json
      description: Validates a QR code payload for integrity (no authentication required)
      parameters:
      - description: QR payload to validate
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.QrValidationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Validation result
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Validation failed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Validate QR code payload
      tags:
      - public
...
  /batteries/{bpan}/qr:
    get:
      description: Generates a QR code PNG image containing the battery's BPAN payload
      parameters:
      - description: Battery PAN
        in: path
        name: bpan
        required: true
        type: string
      produces:
      - image/png
      responses:
        "200":
          description: QR Code PNG image
          schema:
            type: file
        "400":
          description: Invalid BPAN
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: QR generation failed
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Generate QR code for a battery
      tags:
      - qr
...
  /batteries/scan:
    post:
      consumes:
      - application/json
      description: Decodes a QR code payload and returns battery information
      parameters:
      - description: QR payload
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Validation failed
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Scan and decode a QR code
      tags:
      - qr
...
swagger: "2.0"